bluetooth: Don't crash when the local adapter is not set
In devices where bluetooth is not working (like the emulator or devices
in development), the bluetooth Settings crashes when closing the activity.

Most of the code checks for the nullability of the local adapter except
for this particular piece. This is usually not a worry for production
devices and I wouldn't usually worry about this but since the rest of the
code performs the nullability check it seems silly to crash onPause.

Change-Id: Ie0cef58061fdc585f3b956104a5407b948005775
Reviewed-on: http://review.blissroms.com/4261
Reviewed-by: Jackeagle <jackeagle102@gmail.com>
Tested-by: Jackeagle <jackeagle102@gmail.com>